# api/openapi.yaml
---
openapi: 3.0.3
info:
  title: Sample API
  description: API description in Markdown.
  version: 1.0.0
servers:
  - url: 'https://api.example.com/api/v1'
paths:
  /pages:
    get:
      operationId: getPages
      summary: Get all pages
      parameters:
        - in: query
          name: tags
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: All pages data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pages'
        default:
          $ref: '#/components/responses/undefinedError'
    post:
      operationId: addPage
      summary: Add new page
      parameters:
        - in: query
          name: url
          schema:
            type: string
        - in: query
          name: description
          schema:
            type: string
        - in: query
          name: formats
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/format'
        - in: query
          name: tags
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                description:
                  type: string
                formats:
                  type: array
                  items:
                    $ref: '#/components/schemas/format'
                tags:
                  type: array
                  items:
                    type: string
                collection_id:
                  type: string
                  format: uuid
                  nullable: true
                depth:
                  type: integer
                  default: 0
                  description: "Note: link depth > 1 might cause large archives"
                headers:
                  type: object
                  additionalProperties:
                    type: string
                cookies:
                  type: object
                  additionalProperties:
                    type: string
              required:
                - url
      responses:
        201:
          description: Page added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  field:
                    type: string
                    nullable: false
                  error:
                    type: string
                    nullable: false
                required:
                  - error
                  - field
        default:
          $ref: '#/components/responses/undefinedError'

  /collections:
    get:
      operationId: getCollections
      summary: Get all collections
      responses:
        200:
          description: All collections data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/collection'
        default:
          $ref: '#/components/responses/undefinedError'
    post:
      operationId: addCollection
      summary: Add new collection
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required:
                - name
      responses:
        201:
          description: Collection added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/collection'
        default:
          $ref: '#/components/responses/undefinedError'

  /collections/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getCollection
      description: Get collection details
      responses:
        200:
          description: Collection data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/collection'
        404:
          description: Collection not found
        default:
          $ref: '#/components/responses/undefinedError'
    delete:
      operationId: deleteCollection
      description: Delete a collection
      responses:
        204:
          description: Collection deleted
        404:
          description: Collection not found
        default:
          $ref: '#/components/responses/undefinedError'

  /pages/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getPage
      description: Get page details
      responses:
        200:
          description: Page data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pageWithResults'
        404:
          description: Page not found
        default:
          $ref: '#/components/responses/undefinedError'
    patch:
      operationId: updatePage
      description: Edit page details (title, description, tags)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: Page updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page'
        404:
          description: Page not found
        default:
          $ref: '#/components/responses/undefinedError'
    delete:
      operationId: deletePage
      description: Delete a page and its files
      responses:
        204:
          description: Page deleted
        404:
          description: Page not found
        default:
          $ref: '#/components/responses/undefinedError'

  /pages/{id}/file/{file_id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: file_id
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getFile
      description: Get file content
      responses:
        200:
          description: File content
          content:
            application/pdf: {}
            text/plain:
              schema:
                type: string
            text/html:
              schema:
                type: string
        404:
          description: Page of file not found
        default:
          $ref: '#/components/responses/undefinedError'

components:
  responses:
    undefinedError:
      description: Undefined Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  schemas:
    format:
      type: string
      enum:
        - all
        - pdf
        - single_file
        - headers
        - markdown
    error:
      type: object
      properties:
        message:
          type: string
        localized:
          type: string
      required:
        - message
    pages:
      type: array
      items:
        $ref: '#/components/schemas/page'
    page:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        created:
          type: string
          format: date-time
        collection_id:
          type: string
          format: uuid
          nullable: true
        depth:
          type: integer
        formats:
          type: array
          items:
            $ref: '#/components/schemas/format'
        tags:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/status'
        meta:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            error:
              type: string
          required:
            - title
            - description
      required:
        - id
        - url
        - formats
        - status
        - created
        - meta
        - depth
    collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created:
          type: string
          format: date-time
      required:
        - id
        - name
        - created
    result:
      type: object
      properties:
        format:
          $ref: '#/components/schemas/format'
        error:
          type: string
        files:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              mimetype:
                type: string
              size:
                type: integer
                format: int64
            required:
              - id
              - name
              - mimetype
              - size
      required:
        - format
        - files
    pageWithResults:
      allOf:
        - $ref: '#/components/schemas/page'
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/result'
          required:
            - results
    status:
      type: string
      enum:
        - new
        - processing
        - done
        - failed
        - with_errors